+++ scripts/release/release.sh
@@ -54,10 +54,10 @@ push_and_publish() {
   push_release_branch_and_tag
 
   log "Publishing package on npm"
-  npm publish --tag "$dist_tag" --otp "$(prompt_otp)"
+  npm publish --tag "$dist_tag" --otp "$(prompt_otp)" --access=public
 
   if [[ "$dist_tag" == "latest" ]]; then
-    npm dist-tag rm --otp "$(prompt_otp)" openzeppelin-test-helpers next
+    npm dist-tag rm --otp "$(prompt_otp)" @openzeppelin/test-helpers next
   fi
 }