+++ src/expectRevert.js
@@ -36,7 +36,7 @@ async function checkRevertReasonSupport (provider) {
 
     const warn = function (msg) {
       console.log(`\
-${colors.white.bgBlack('openzeppelin-test-helpers')} ${colors.black.bgYellow('WARN')} expectRevert: ${msg}`
+${colors.white.bgBlack('@openzeppelin/test-helpers')} ${colors.black.bgYellow('WARN')} expectRevert: ${msg}`
       );
     };