+++ test-integration/truffle-migration/migrations/3_test_singleton.js
@@ -1,5 +1,5 @@
 module.exports = async function(deployer, network, accounts) {
-  const { singletons } = require('openzeppelin-test-helpers');
+  const { singletons } = require('@openzeppelin/test-helpers');
 
   await singletons.ERC1820Registry(accounts[0]);